+++ desc/ae/tagger/BricsTaggerAE.xml
@@ -2,7 +2,7 @@
 <analysisEngineDescription xmlns="http://uima.apache.org/resourceSpecifier">
   <frameworkImplementation>org.apache.uima.java</frameworkImplementation>
   <primitive>true</primitive>
-  <annotatorImplementationName>de.berlin.hu.uima.ae.tagger.linnaeus.BricsTagger</annotatorImplementationName>
+  <annotatorImplementationName>de.berlin.hu.uima.ae.tagger.brics.BricsTagger</annotatorImplementationName>
   <analysisEngineMetaData>
     <name>bricsTaggerAEDescriptor</name>
     <description/>
